# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports Student Digital Report Card API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Student {
  id: number;
  name: string;
  className: string;
  birthDate: string;
  height: number;
  weight: number;
  createdAt: string;
}

export interface CreateStudentBody {
  name: string;
  className: string;
  birthDate: string;
  height: number;
  weight: number;
}

export interface UpdateStudentBody {
  name?: string;
  className?: string;
  birthDate?: string;
  height?: number;
  weight?: number;
}

export type PerformanceMetricSport =
  (typeof PerformanceMetricSport)[keyof typeof PerformanceMetricSport];

export const PerformanceMetricSport = {
  voleybol: "voleybol",
  basketbol: "basketbol",
  futbol: "futbol",
} as const;

export interface PerformanceMetric {
  id: number;
  studentId: number;
  sport: PerformanceMetricSport;
  speed: number;
  endurance: number;
  strength: number;
  agility: number;
  technique: number;
  teamwork: number;
  overallScore: number;
  /** @nullable */
  notes?: string | null;
  evaluationDate: string;
  createdAt: string;
}

export type CreateMetricBodySport =
  (typeof CreateMetricBodySport)[keyof typeof CreateMetricBodySport];

export const CreateMetricBodySport = {
  voleybol: "voleybol",
  basketbol: "basketbol",
  futbol: "futbol",
} as const;

export interface CreateMetricBody {
  studentId: number;
  sport: CreateMetricBodySport;
  speed: number;
  endurance: number;
  strength: number;
  agility: number;
  technique: number;
  teamwork: number;
  /** @nullable */
  notes?: string | null;
  evaluationDate: string;
}

export interface UpdateMetricBody {
  speed?: number;
  endurance?: number;
  strength?: number;
  agility?: number;
  technique?: number;
  teamwork?: number;
  /** @nullable */
  notes?: string | null;
  evaluationDate?: string;
}

export interface SportAverage {
  sport: string;
  averageScore: number;
  metricCount: number;
}

export interface ReportCard {
  student: Student;
  metrics: PerformanceMetric[];
  sportAverages: SportAverage[];
  overallAverage: number;
}

export interface RankingEntry {
  rank: number;
  studentId: number;
  studentName: string;
  className: string;
  averageScore: number;
  metricCount: number;
}

export interface SportBreakdown {
  sport: string;
  count: number;
  averageScore: number;
}

export interface TopPerformer {
  studentId: number;
  studentName: string;
  sport: string;
  averageScore: number;
}

export interface DashboardSummary {
  totalStudents: number;
  totalMetrics: number;
  sportBreakdown: SportBreakdown[];
  topPerformers: TopPerformer[];
}

export type ActivityEntryType =
  (typeof ActivityEntryType)[keyof typeof ActivityEntryType];

export const ActivityEntryType = {
  student_added: "student_added",
  metric_added: "metric_added",
  student_updated: "student_updated",
  metric_updated: "metric_updated",
} as const;

export interface ActivityEntry {
  id: number;
  type: ActivityEntryType;
  description: string;
  timestamp: string;
}

export interface ClassStat {
  className: string;
  studentCount: number;
  averageHeight: number;
  averageWeight: number;
  averageScore: number;
}

export type ListStudentsParams = {
  classFilter?: string;
  search?: string;
};

export type ListMetricsParams = {
  studentId?: number;
  sport?: string;
};