# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports Student Digital Report Card API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all students
 */
export const ListStudentsQueryParams = zod.object({
  classFilter: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
});

export const ListStudentsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  className: zod.string(),
  birthDate: zod.coerce.date(),
  height: zod.number(),
  weight: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListStudentsResponse = zod.array(ListStudentsResponseItem);

/**
 * @summary Add a new student
 */
export const CreateStudentBody = zod.object({
  name: zod.string(),
  className: zod.string(),
  birthDate: zod.coerce.date(),
  height: zod.number(),
  weight: zod.number(),
});

/**
 * @summary Get student by ID
 */
export const GetStudentParams = zod.object({
  id: zod.coerce.number(),
});

export const GetStudentResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  className: zod.string(),
  birthDate: zod.coerce.date(),
  height: zod.number(),
  weight: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update student info
 */
export const UpdateStudentParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateStudentBody = zod.object({
  name: zod.string().optional(),
  className: zod.string().optional(),
  birthDate: zod.coerce.date().optional(),
  height: zod.number().optional(),
  weight: zod.number().optional(),
});

export const UpdateStudentResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  className: zod.string(),
  birthDate: zod.coerce.date(),
  height: zod.number(),
  weight: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a student
 */
export const DeleteStudentParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get full report card for a student
 */
export const GetStudentReportCardParams = zod.object({
  id: zod.coerce.number(),
});

export const GetStudentReportCardResponse = zod.object({
  student: zod.object({
    id: zod.number(),
    name: zod.string(),
    className: zod.string(),
    birthDate: zod.coerce.date(),
    height: zod.number(),
    weight: zod.number(),
    createdAt: zod.coerce.date(),
  }),
  metrics: zod.array(
    zod.object({
      id: zod.number(),
      studentId: zod.number(),
      sport: zod.enum(["voleybol", "basketbol", "futbol"]),
      speed: zod.number(),
      endurance: zod.number(),
      strength: zod.number(),
      agility: zod.number(),
      technique: zod.number(),
      teamwork: zod.number(),
      overallScore: zod.number(),
      notes: zod.string().nullish(),
      evaluationDate: zod.coerce.date(),
      createdAt: zod.coerce.date(),
    }),
  ),
  sportAverages: zod.array(
    zod.object({
      sport: zod.string(),
      averageScore: zod.number(),
      metricCount: zod.number(),
    }),
  ),
  overallAverage: zod.number(),
});

/**
 * @summary List performance metrics
 */
export const ListMetricsQueryParams = zod.object({
  studentId: zod.coerce.number().optional(),
  sport: zod.coerce.string().optional(),
});

export const ListMetricsResponseItem = zod.object({
  id: zod.number(),
  studentId: zod.number(),
  sport: zod.enum(["voleybol", "basketbol", "futbol"]),
  speed: zod.number(),
  endurance: zod.number(),
  strength: zod.number(),
  agility: zod.number(),
  technique: zod.number(),
  teamwork: zod.number(),
  overallScore: zod.number(),
  notes: zod.string().nullish(),
  evaluationDate: zod.coerce.date(),
  createdAt: zod.coerce.date(),
});
export const ListMetricsResponse = zod.array(ListMetricsResponseItem);

/**
 * @summary Add a performance metric
 */
export const CreateMetricBody = zod.object({
  studentId: zod.number(),
  sport: zod.enum(["voleybol", "basketbol", "futbol"]),
  speed: zod.number(),
  endurance: zod.number(),
  strength: zod.number(),
  agility: zod.number(),
  technique: zod.number(),
  teamwork: zod.number(),
  notes: zod.string().nullish(),
  evaluationDate: zod.coerce.date(),
});

/**
 * @summary Update a metric
 */
export const UpdateMetricParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateMetricBody = zod.object({
  speed: zod.number().optional(),
  endurance: zod.number().optional(),
  strength: zod.number().optional(),
  agility: zod.number().optional(),
  technique: zod.number().optional(),
  teamwork: zod.number().optional(),
  notes: zod.string().nullish(),
  evaluationDate: zod.coerce.date().optional(),
});

export const UpdateMetricResponse = zod.object({
  id: zod.number(),
  studentId: zod.number(),
  sport: zod.enum(["voleybol", "basketbol", "futbol"]),
  speed: zod.number(),
  endurance: zod.number(),
  strength: zod.number(),
  agility: zod.number(),
  technique: zod.number(),
  teamwork: zod.number(),
  overallScore: zod.number(),
  notes: zod.string().nullish(),
  evaluationDate: zod.coerce.date(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a metric
 */
export const DeleteMetricParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get rankings for a sport
 */
export const GetRankingsParams = zod.object({
  sport: zod.enum(["voleybol", "basketbol", "futbol"]),
});

export const GetRankingsResponseItem = zod.object({
  rank: zod.number(),
  studentId: zod.number(),
  studentName: zod.string(),
  className: zod.string(),
  averageScore: zod.number(),
  metricCount: zod.number(),
});
export const GetRankingsResponse = zod.array(GetRankingsResponseItem);

/**
 * @summary Get dashboard overview stats
 */
export const GetDashboardSummaryResponse = zod.object({
  totalStudents: zod.number(),
  totalMetrics: zod.number(),
  sportBreakdown: zod.array(
    zod.object({
      sport: zod.string(),
      count: zod.number(),
      averageScore: zod.number(),
    }),
  ),
  topPerformers: zod.array(
    zod.object({
      studentId: zod.number(),
      studentName: zod.string(),
      sport: zod.string(),
      averageScore: zod.number(),
    }),
  ),
});

/**
 * @summary Get recent activity feed
 */
export const GetRecentActivityResponseItem = zod.object({
  id: zod.number(),
  type: zod.enum([
    "student_added",
    "metric_added",
    "student_updated",
    "metric_updated",
  ]),
  description: zod.string(),
  timestamp: zod.coerce.date(),
});
export const GetRecentActivityResponse = zod.array(
  GetRecentActivityResponseItem,
);

/**
 * @summary Get student counts and averages per class
 */
export const GetClassStatsResponseItem = zod.object({
  className: zod.string(),
  studentCount: zod.number(),
  averageHeight: zod.number(),
  averageWeight: zod.number(),
  averageScore: zod.number(),
});
export const GetClassStatsResponse = zod.array(GetClassStatsResponseItem);