# lib/api-zod/src/generated/types/activityEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports Student Digital Report Card API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEntryType } from "./activityEntryType";

export interface ActivityEntry {
  id: number;
  type: ActivityEntryType;
  description: string;
  timestamp: Date;
}