# lib/api-zod/src/generated/types/classStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports Student Digital Report Card API
 * OpenAPI spec version: 0.1.0
 */

export interface ClassStat {
  className: string;
  studentCount: number;
  averageHeight: number;
  averageWeight: number;
  averageScore: number;
}