# lib/api-zod/src/generated/types/createMetricBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports Student Digital Report Card API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateMetricBodySport } from "./createMetricBodySport";

export interface CreateMetricBody {
  studentId: number;
  sport: CreateMetricBodySport;
  speed: number;
  endurance: number;
  strength: number;
  agility: number;
  technique: number;
  teamwork: number;
  /** @nullable */
  notes?: string | null;
  evaluationDate: Date;
}