# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports Student Digital Report Card API
 * OpenAPI spec version: 0.1.0
 */
import type { SportBreakdown } from "./sportBreakdown";
import type { TopPerformer } from "./topPerformer";

export interface DashboardSummary {
  totalStudents: number;
  totalMetrics: number;
  sportBreakdown: SportBreakdown[];
  topPerformers: TopPerformer[];
}