# lib/api-zod/src/generated/types/performanceMetric.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports Student Digital Report Card API
 * OpenAPI spec version: 0.1.0
 */
import type { PerformanceMetricSport } from "./performanceMetricSport";

export interface PerformanceMetric {
  id: number;
  studentId: number;
  sport: PerformanceMetricSport;
  speed: number;
  endurance: number;
  strength: number;
  agility: number;
  technique: number;
  teamwork: number;
  overallScore: number;
  /** @nullable */
  notes?: string | null;
  evaluationDate: Date;
  createdAt: Date;
}