# lib/api-zod/src/generated/types/reportCard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports Student Digital Report Card API
 * OpenAPI spec version: 0.1.0
 */
import type { PerformanceMetric } from "./performanceMetric";
import type { SportAverage } from "./sportAverage";
import type { Student } from "./student";

export interface ReportCard {
  student: Student;
  metrics: PerformanceMetric[];
  sportAverages: SportAverage[];
  overallAverage: number;
}