# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports Student Digital Report Card API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  className: string;
  birthDate: Date;
  height: number;
  weight: number;
  createdAt: Date;
}