# lib/api-zod/src/generated/types/updateStudentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports Student Digital Report Card API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateStudentBody {
  name?: string;
  className?: string;
  birthDate?: Date;
  height?: number;
  weight?: number;
}